fix: 32-bit build and macOS clang CI failures
- durable_task.c: use (void *)(intptr_t) instead of (void *)(long long)
  to avoid int-to-pointer-cast warning on 32-bit builds where pointers
  are 4 bytes but long long is 8 bytes
- unit/Makefile: add -Wno-character-conversion for clang to suppress
  googletest 1.17 header warning in gtest-printers.h

+++ src/durable_task.c
@@ -277,7 +277,7 @@ bool durabilitySignalModifiedKey(struct client *c, struct serverDb *db, struct s
 bool durabilitySignalFlushedDb(int dbid) {
     /* Defer flush invalidation messages until the durability providers acknowledge. */
     return durabilityRegisterDeferredTask(DURABLE_FLUSH_INVALIDATION_TASK,
-                                          (void *)(long long)(dbid == -1));
+                                          (void *)(intptr_t)(dbid == -1));
 }
 
 /*================================= Task Execution =========================== */
